refactor : 의미있는 패키지명으로 수정
